// ui/theDevBlog/src/app/model/update-post.model.ts
export interface UpdatePostRequest {
  title: string | undefined;
  content: string | undefined;
  summary: string | undefined;
  author: string | undefined;
  urlHandle: string | undefined;
  visible: boolean | undefined;
  publishedDate: Date | undefined;
  updatedDate: Date | undefined;
  featuredImageUrl: string | undefined;
}